# Finsecure/api-docs.json
{"openapi":"3.1.0","info":{"title":"finSecure API","description":"Finance dashboard backend","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Authentication","description":"User registration and JWT token acquisition"},{"name":"Records","description":"Financial record management (CRUD & Filtering)"},{"name":"Admin","description":"System-wide user management and role assignment (ADMIN only)"},{"name":"Dashboard","description":"Analytics and summary endpoints"}],"paths":{"/finsecure/records/{id}":{"get":{"tags":["Records"],"summary":"Get record details by ID (ADMIN, ANALYST)","description":"Fetches a single financial record using its unique UUID. Accessible only to high-privilege roles.","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RecordResponse"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["Records"],"summary":"Update an existing record (ADMIN only)","description":"Modifies the details of a specific financial record. Changes to amounts or categories are restricted to the ADMIN role.","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RecordResponse"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Records"],"summary":"Delete a record (ADMIN only)","description":"Permanently removes a record from the database. This action is destructive and limited to the ADMIN role.","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/records":{"get":{"tags":["Records"],"summary":"List all records with dynamic filtering (ADMIN, ANALYST)","description":"Retrieves a paginated list of financial records. Allows filtering by category, record type, and date range. Provides the system-wide oversight required for administrative and analytical review.","operationId":"getAll","parameters":[{"name":"category","in":"query","required":false,"schema":{"type":"string","enum":["FOOD","RENT","SALARY","ENTERTAINMENT","UTILITIES","INVESTMENT","OTHER"]}},{"name":"recordType","in":"query","required":false,"schema":{"type":"string","enum":["EXPENSE","INCOME"]}},{"name":"from","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"to","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageRecordResponse"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Records"],"summary":"Create a new financial record (ADMIN only)","description":"Saves a new income or expense record to the system. Restricted to users with the ADMIN role.","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RecordResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/auth/register":{"post":{"tags":["Authentication"],"summary":"Register a new user account","description":"Creates a new user in the system. By default, all new users are assigned the 'VIEWER' role.","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/auth/login":{"post":{"tags":["Authentication"],"summary":"Authenticate user and receive JWT","description":"Verifies credentials and returns a stateless JWT token to be used in the 'Authorization' header for protected endpoints.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/admin/users/{id}/status":{"patch":{"tags":["Admin"],"summary":"Activate or Deactivate user account","description":"Toggles a user's access to the system. Deactivated users are blocked from logging in via the security filter.","operationId":"updateStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"isActive","in":"query","required":true,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/admin/users/{id}/role":{"patch":{"tags":["Admin"],"summary":"Promote or Demote user role","description":"Assigns a new role (VIEWER, ANALYST, ADMIN) to an existing user to change their system permissions.","operationId":"updateRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"role","in":"query","required":true,"schema":{"type":"string","enum":["ANALYST","VIEWER","ADMIN"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/dashboard/my/trends":{"get":{"tags":["Dashboard"],"summary":"My monthly trends — last 6 months","description":"Returns month-by-month income vs expense trends for the authenticated user.","operationId":"getMyTrends","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TrendResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/dashboard/my/totals":{"get":{"tags":["Dashboard"],"summary":"My totals — income, expenses, net balance","description":"Returns financial totals scoped to the authenticated user's own records.","operationId":"getMyTotals","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TotalsResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/dashboard/my/recent":{"get":{"tags":["Dashboard"],"summary":"My recent activity — last 10 transactions ","description":"Returns the 10 most recent records for the authenticated user.","operationId":"getMyRecentActivity","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RecordResponse"}}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/dashboard/my/categories":{"get":{"tags":["Dashboard"],"summary":"My category breakdown — income and expenses by category","description":"Returns category-wise totals scoped to the authenticated user.","operationId":"getMyCategorySummary","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategorySummaryResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/dashboard/global/trends":{"get":{"tags":["Dashboard"],"summary":"Global monthly trends — configurable window (ANALYST, ADMIN)","description":"Platform-wide month-by-month trends. Use ?months=6 (default) to control the lookback window.","operationId":"getGlobalTrends","parameters":[{"name":"months","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":6}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TrendResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/dashboard/global/totals":{"get":{"tags":["Dashboard"],"summary":"Global totals — platform-wide income, expenses, net (ANALYST, ADMIN)","description":"Aggregates all records across all users. Allows Analyst to view the system's financial health.","operationId":"getGlobalTotals","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TotalsResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/dashboard/global/summary":{"get":{"tags":["Dashboard"],"summary":"Global full summary — complete platform-wide overview (ADMIN only)","description":"One-shot endpoint combining totals, categories, trends and recent activity across all users.","operationId":"getGlobalSummary","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardSummaryResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/dashboard/global/recent":{"get":{"tags":["Dashboard"],"summary":"Global recent activity — last 10 transactions across all users (ANALYST, ADMIN)","description":"Returns the 10 most recent records platform-wide, regardless of which user created them.","operationId":"getGlobalRecentActivity","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RecordResponse"}}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/dashboard/global/categories":{"get":{"tags":["Dashboard"],"summary":"Global category breakdown — across all users (ANALYST, ADMIN)","description":"Category-wise income and expense totals across the entire platform.","operationId":"getGlobalCategorySummary","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CategorySummaryResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/dashboard/global/basic/summary":{"get":{"tags":["Dashboard"],"summary":"Global full summary — Basic summary for the viewer and analyst","description":"One-shot endpoint combining totals, categories, trends and recent activity across all users.","operationId":"getBasicSummary","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BasicDashboardSummaryResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/admin/users":{"get":{"tags":["Admin"],"summary":"List all registered users","description":"Retrieves a list of all users in the system. Can be filtered by active/inactive status using the 'isActive' parameter.","operationId":"getUsers","parameters":[{"name":"isActive","in":"query","required":false,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/admin/users/{id}":{"get":{"tags":["Admin"],"summary":"Get user details by ID","description":"Fetches the full profile of a specific user, excluding sensitive credentials like passwords.","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Admin"],"summary":"Permanently delete a user","description":"Hard delete of a user account and all associated records from the database. This action is irreversible.","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/finsecure/admin/users/delete/{id}":{"delete":{"tags":["Admin"],"summary":"Soft delete a user","description":"Marks a user as deleted without removing their historical data from the database.","operationId":"softDelete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"RecordRequest":{"type":"object","properties":{"amount":{"type":"number","minimum":0.01},"category":{"type":"string","enum":["FOOD","RENT","SALARY","ENTERTAINMENT","UTILITIES","INVESTMENT","OTHER"]},"recordType":{"type":"string","enum":["EXPENSE","INCOME"]},"note":{"type":"string","minLength":1},"transactionDate":{"type":"string","format":"date"}},"required":["amount","category","note","recordType"]},"RecordResponse":{"type":"object","properties":{"recordId":{"type":"string","format":"uuid"},"amount":{"type":"number"},"category":{"type":"string","enum":["FOOD","RENT","SALARY","ENTERTAINMENT","UTILITIES","INVESTMENT","OTHER"]},"recordType":{"type":"string","enum":["EXPENSE","INCOME"]},"note":{"type":"string"},"transactionDate":{"type":"string","format":"date"},"createdAt":{"type":"string","format":"date-time"}}},"UserRegisterRequest":{"type":"object","properties":{"username":{"type":"string","minLength":1},"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":2147483647,"minLength":6}},"required":["email","password","username"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"message":{"type":"string"}}},"PageRecordResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/RecordResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"paged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"MonthlyTrend":{"type":"object","properties":{"month":{"type":"string"},"income":{"type":"number"},"expenses":{"type":"number"},"net":{"type":"number"}}},"TrendResponse":{"type":"object","properties":{"monthlyTrends":{"type":"array","items":{"$ref":"#/components/schemas/MonthlyTrend"}}}},"TotalsResponse":{"type":"object","properties":{"totalIncome":{"type":"number"},"totalExpenses":{"type":"number"},"netBalance":{"type":"number"}}},"CategorySummaryResponse":{"type":"object","properties":{"expensesByCategory":{"type":"object","additionalProperties":{"type":"number"}},"incomeByCategory":{"type":"object","additionalProperties":{"type":"number"}}}},"DashboardSummaryResponse":{"type":"object","properties":{"totalIncome":{"type":"number"},"totalExpense":{"type":"number"},"netBalance":{"type":"number"},"expensesByCategory":{"type":"object","additionalProperties":{"type":"number"}},"incomeByCategory":{"type":"object","additionalProperties":{"type":"number"}},"monthlyTrends":{"type":"array","items":{"$ref":"#/components/schemas/MonthlyTrend"}},"recentActivity":{"type":"array","items":{"$ref":"#/components/schemas/RecordResponse"}}}},"BasicDashboardSummaryResponse":{"type":"object","properties":{"totalIncome":{"type":"number"},"totalExpense":{"type":"number"},"netBalance":{"type":"number"},"recentActivity":{"type":"array","items":{"$ref":"#/components/schemas/RecordResponse"}}}},"UserResponse":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"username":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["ANALYST","VIEWER","ADMIN"]},"isActive":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}